requirements.txt
PyQt6==6.4.2
PyYAML==6.0.1